fix: make nouns singular
Close https://github.com/openfga/cli/issues/50
