+++ cmd/store/get.go
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package stores
+package store
 
 import (
 	"context"
@@ -22,30 +22,29 @@ import (
 	"os"
 
 	"github.com/openfga/cli/lib/cmd-utils"
+	"github.com/openfga/cli/lib/fga"
 	"github.com/openfga/go-sdk/client"
 	"github.com/spf13/cobra"
 )
 
-func create(fgaClient client.SdkClient, storeName string) (string, error) {
-	body := client.ClientCreateStoreRequest{Name: storeName}
-
-	store, err := fgaClient.CreateStore(context.Background()).Body(body).Execute()
+func getStore(clientConfig fga.ClientConfig, fgaClient client.SdkClient) (string, error) {
+	store, err := fgaClient.GetStore(context.Background()).Execute()
 	if err != nil {
-		return "", fmt.Errorf("failed to create store %v due to %w", storeName, err)
+		return "", fmt.Errorf("failed to get store %v due to %w", clientConfig.StoreID, err)
 	}
 
 	storeJSON, err := json.Marshal(store)
 	if err != nil {
-		return "", fmt.Errorf("store %v created, but failed to be printed due to %w", *store.Id, err)
+		return "", fmt.Errorf("failed to get store due to %w", err)
 	}
 
 	return string(storeJSON), nil
 }
 
-// createCmd represents the store create command.
-var createCmd = &cobra.Command{
-	Use:   "create",
-	Short: "Create and initialize a store.",
+// getCmd represents the get command.
+var getCmd = &cobra.Command{
+	Use:   "get",
+	Short: "Get store",
 	Long:  ``,
 	RunE: func(cmd *cobra.Command, args []string) error {
 		clientConfig := cmdutils.GetClientConfig(cmd)
@@ -53,11 +52,12 @@ var createCmd = &cobra.Command{
 		if err != nil {
 			return fmt.Errorf("failed to initialize FGA Client due to %w", err)
 		}
-		storeName, _ := cmd.Flags().GetString("name")
-		output, err := create(fgaClient, storeName)
+
+		output, err := getStore(clientConfig, fgaClient)
 		if err != nil {
 			return err
 		}
+
 		fmt.Print(output)
 
 		return nil
@@ -65,8 +65,8 @@ var createCmd = &cobra.Command{
 }
 
 func init() {
-	createCmd.Flags().String("name", "", "Store Name")
-	err := createCmd.MarkFlagRequired("name")
+	getCmd.Flags().String("store-id", "", "Store ID")
+	err := getCmd.MarkFlagRequired("store-id")
 	if err != nil { //nolint:wsl
 		fmt.Print(err)
 		os.Exit(1)
