+++ cmd/store/store.go
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package stores
+package store
 
 import (
 	"github.com/spf13/cobra"
@@ -21,7 +21,7 @@ import (
 
 // StoresCmd represents the store command.
 var StoresCmd = &cobra.Command{
-	Use:   "stores",
+	Use:   "store",
 	Short: "Create, Get, Delete and List OpenFGA Stores",
 }
