+++ cmd/tuple/changes_test.go
@@ -1,4 +1,4 @@
-package tuples
+package tuple
 
 import (
 	"context"
