+++ cmd/tuple/tuple.go
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package models
+package tuple
 
 import (
 	"fmt"
@@ -22,19 +22,20 @@ import (
 	"github.com/spf13/cobra"
 )
 
-// ModelsCmd represents the store command.
-var ModelsCmd = &cobra.Command{
-	Use:   "models",
-	Short: "Write, Read and List authorization models in a store",
+// TupleCmd represents the tuple command.
+var TupleCmd = &cobra.Command{
+	Use:   "tuple",
+	Short: "Interact with Relationship Tuples in a store.",
 }
 
 func init() {
-	ModelsCmd.AddCommand(writeCmd)
-	ModelsCmd.AddCommand(listCmd)
-	ModelsCmd.AddCommand(getCmd)
+	TupleCmd.AddCommand(writeCmd)
+	TupleCmd.AddCommand(deleteCmd)
+	TupleCmd.AddCommand(readCmd)
+	TupleCmd.AddCommand(changesCmd)
 
-	ModelsCmd.PersistentFlags().String("store-id", "", "Store ID")
-	err := ModelsCmd.MarkPersistentFlagRequired("store-id")
+	TupleCmd.PersistentFlags().String("store-id", "", "Store ID")
+	err := TupleCmd.MarkPersistentFlagRequired("store-id")
 	if err != nil { //nolint:wsl
 		fmt.Print(err)
 		os.Exit(1)
